+++ app/src/main/java/com/groomify/hollavirun/utils/SharedPreferencesHelper.java
@@ -42,8 +42,8 @@ public class SharedPreferencesHelper {
         return settings.getString(AppConstant.PREFS_FB_ID, null);
     }
 
-    public static Integer getSelectedRaceId(Context context){
+    public static Long getSelectedRaceId(Context context){
         SharedPreferences settings = context.getSharedPreferences(AppConstant.PREFS_NAME, 0);
-        return settings.getInt(AppConstant.PREFS_RUN_SELECTED_ID, -1);
+        return settings.getLong(AppConstant.PREFS_RUN_SELECTED_ID, -1);
     }
 }
